requirements.txt

markdown
tqdm
fastapi
uvicorn
qdrant-client
sentence-transformers
python-dotenv
google-generativeai

